+++ tests/specs/cleanup-on-halt.e2e.js
@@ -1,10 +1,11 @@
+import { rumFlags } from '../../tools/testing-server/constants'
 import { testRumRequest } from '../../tools/testing-server/utils/expect-tests'
 
 describe('Memory leaks', () => {
   it('does not occur on ee backlog when RUM flags are 0', async () => {
     await browser.testHandle.scheduleReply('bamServer', {
       test: testRumRequest,
-      body: JSON.stringify({ st: 0, sr: 0, err: 0, ins: 0, spa: 0, loaded: 1 })
+      body: JSON.stringify(rumFlags({ st: 0, sr: 0, err: 0, ins: 0, spa: 0, loaded: 1 }))
     })
     // This test relies on features to call deregisterDrain when not enabled by flags which in turn should clear their backlogs.