+++ tests/specs/session-trace/trace-nodes.e2e.js
@@ -1,3 +1,4 @@
+import { rumFlags } from '../../../tools/testing-server/constants.js'
 import { testBlobTraceRequest, testRumRequest } from '../../../tools/testing-server/utils/expect-tests.js'
 import { stConfig } from '../util/helpers.js'
 import { JSONPath } from 'jsonpath-plus'
@@ -10,7 +11,7 @@ describe('Trace nodes', () => {
 
     await browser.testHandle.scheduleReply('bamServer', {
       test: testRumRequest,
-      body: JSON.stringify({ st: 1, sts: 1, err: 1, ins: 1, spa: 1, sr: 0, loaded: 1 })
+      body: JSON.stringify(rumFlags({ sr: 0 }))
     })
   })
 
@@ -46,7 +47,7 @@ describe('Trace nodes', () => {
     await browser.testHandle.clearScheduledReplies('bamServer')
     await browser.testHandle.scheduleReply('bamServer', {
       test: testRumRequest,
-      body: JSON.stringify({ st: 0, sts: 0, sr: 0, err: 1, ins: 1, spa: 0, loaded: 1 })
+      body: JSON.stringify(rumFlags({ st: 0, sts: 0, sr: 0, spa: 0 }))
     })
 
     const url = await browser.testHandle.assetURL('pagehide.html', stConfig())